Completed Temp Files. Tested.
Improved storage of custom data in resource writers.

+++ src/main/java/jacz/peerengineservice/util/datatransfer/master/DownloadManager.java
@@ -57,7 +57,7 @@ public class DownloadManager {
         return masterResourceStreamer.getMasterPriority();
     }
 
-    public void setPriority(int priority) {
+    public void setPriority(float priority) {
         masterResourceStreamer.setPriority(priority);
     }
